move contents of aida_api into aida
