--- jdaviz/core/tests/test_aida_api.py
@@ -1,169 +1,0 @@
-import pytest
-import re
-import numpy as np
-import astropy.units as u
-from astropy.coordinates import SkyCoord
-from astropy.tests.helper import assert_quantity_allclose
-
-
-def assert_coordinate_close(coord1, coord2, atol=1 * u.arcsec):
-    # check that two coordinates are within some separation tolerance
-    separation = coord1.separation(coord2)
-    assert_quantity_allclose(separation, desired=0*u.arcsec, atol=atol)
-
-
-def assert_angle_close(angle1, angle2, atol=1 * u.arcsec):
-    # check that two angles are within some separation tolerance
-    difference = abs(angle1 - angle2)
-    assert_quantity_allclose(difference, desired=0*u.arcsec, atol=atol)
-
-
-def test_get_viewport_sky(deconfigged_helper, image_hdu_wcs):
-    deconfigged_helper.load(image_hdu_wcs, format='Image')
-    deconfigged_helper.plugins['Orientation'].align_by = 'WCS'
-
-    viewer = deconfigged_helper.viewers['Image']
-    viewport = viewer.get_viewport('sky')
-
-    expected_center = SkyCoord(ra=337.51894337, dec=-20.83208305, unit='deg')
-    expected_fov = 0.00277778 * u.deg
-    expected_image_label = deconfigged_helper._app.data_collection[0].label
-
-    assert_coordinate_close(viewport['center'], expected_center)
-    assert_angle_close(viewport['fov'], expected_fov)
-
-    expected_rotation = 360 * u.deg
-    assert_angle_close(viewport['rotation'], expected_rotation, atol=0.001*u.deg)
-
-    assert viewport['image_label'] == expected_image_label
-
-
-def test_set_viewport_sky(deconfigged_helper, image_hdu_wcs):
-    deconfigged_helper.load(image_hdu_wcs, format='Image')
-    deconfigged_helper.plugins['Orientation'].align_by = 'WCS'
-    viewer = deconfigged_helper.viewers['Image']
-
-    # change the center and fov:
-    new_viewport_settings = dict(
-        center=SkyCoord(ra=337.5, dec=-20.8, unit='deg'),
-        fov=0.01 * u.deg,
-    )
-    viewer.set_viewport(**new_viewport_settings)
-    new_viewport = viewer.get_viewport('sky')
-
-    assert_coordinate_close(new_viewport['center'], new_viewport_settings['center'])
-
-    assert_angle_close(new_viewport['fov'], new_viewport_settings['fov'], atol=1 * u.arcsec)
-
-    with pytest.raises(ValueError, match='The AID API supports `center` arguments as'):
-        viewer.set_viewport(center=u.Quantity([0, 1]))
-
-
-def test_set_viewport_sky_rotation(deconfigged_helper, image_hdu_wcs):
-    deconfigged_helper.load(image_hdu_wcs, format='Image')
-    deconfigged_helper.plugins['Orientation'].align_by = 'WCS'
-    viewer = deconfigged_helper.viewers['Image']
-
-    # change only the rotation:
-    new_viewport_settings = dict(
-        rotation=90 * u.deg
-    )
-    viewer.set_viewport(**new_viewport_settings)
-    new_viewport = viewer.get_viewport('sky')
-
-    assert_angle_close(
-        new_viewport['rotation'],
-        new_viewport_settings['rotation'],
-        atol=0.001*u.deg
-    )
-
-
-def test_get_viewport_external_update(deconfigged_helper, image_hdu_wcs):
-    deconfigged_helper.load(image_hdu_wcs, format='Image')
-    deconfigged_helper.plugins['Orientation'].align_by = 'WCS'
-    viewer = deconfigged_helper.viewers['Image']
-
-    # update the image viewer rotation using Orientation plugin
-    orientation = deconfigged_helper._app._jdaviz_helper.plugins.get('Orientation', None)
-    orientation.add_orientation(
-        east_left=True,
-        set_on_create=True,
-        label='test',
-        rotation_angle=90.0
-    )
-
-    # assert: check that aida get viewport gets the correct rotation
-    new_viewport = viewer.get_viewport('sky')
-    assert_angle_close(
-        new_viewport['rotation'],
-        90.0*u.deg,
-        atol=0.001*u.deg
-    )
-
-
-def test_set_viewport_pixel(deconfigged_helper, image_hdu_wcs):
-    deconfigged_helper.load(image_hdu_wcs, format='Image')
-    deconfigged_helper.plugins['Orientation'].align_by = 'WCS'
-
-    viewer = deconfigged_helper.viewers['Image']
-
-    # set the center and fov:
-    new_viewport_settings = dict(
-        center=[5, 5],
-        fov=10
-    )
-    viewer.set_viewport(**new_viewport_settings)
-    new_viewport = viewer.get_viewport('pixel')
-
-    np.testing.assert_allclose(new_viewport['center'], new_viewport_settings['center'])
-
-    np.testing.assert_allclose(new_viewport['fov'], new_viewport_settings['fov'], atol=1e-4)
-
-
-def test_set_rotation_not_aligned(deconfigged_helper, image_hdu_wcs):
-    deconfigged_helper.load(image_hdu_wcs, format='Image')
-    viewer = deconfigged_helper.viewers['Image']
-
-    # change only the rotation:
-    new_viewport_settings = dict(
-        rotation=90 * u.deg
-    )
-    with pytest.raises(ValueError, match=re.escape("The viewer must be aligned by WCS to use `set_rotation`.")):  # noqa
-        viewer.set_viewport(**new_viewport_settings)
-
-
-def test_get_fov_no_wcs(deconfigged_helper, image_hdu_nowcs):
-    deconfigged_helper.load(image_hdu_nowcs, format='Image')
-    viewer = deconfigged_helper.viewers['Image']
-
-    with pytest.raises(ValueError, match=re.escape("The image must have valid WCS to return `fov` in `sky`.")):  # noqa
-        viewer.get_viewport()
-
-
-def test_get_viewport_no_wcs(deconfigged_helper):
-    image_no_wcs = np.random.normal(size=(50, 50))
-    deconfigged_helper.load(image_no_wcs, loader='object', format='Image')
-
-    viewer = deconfigged_helper.viewers['Image']
-    viewport = viewer.get_viewport()
-
-    assert viewport['center'] == (24.5, 24.5)
-    assert viewport['fov'] == 50.0
-    assert viewport['rotation'] is None
-    assert viewport['projection'] is None
-    assert viewport['image_label'] == 'Image'
-
-
-def test_set_viewport_no_wcs(deconfigged_helper):
-    image_no_wcs = np.random.normal(size=(50, 50))
-    deconfigged_helper.load(image_no_wcs, loader='object', format='Image')
-
-    viewer = deconfigged_helper.viewers['Image']
-    viewer.set_viewport(center=(0, 0), fov=10)
-    viewport = viewer.get_viewport()
-
-    assert viewport['center'] == (0, 0)
-    assert viewport['fov'] == 10
-    assert viewport['rotation'] is None
-    assert viewport['projection'] is None
-    assert viewport['image_label'] == 'Image'
